core/local/chokidar: Extract ./initial_scan from ./watcher
